pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "TSSPDCL Voice Agent - AI-powered customer care system for complaint registration and status checking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "pyodbc>=5.0.0",
    "asyncio>=3.4.3",
]
